Create KeyboardControl.py
Consider the action of opening a new tab in a browser, enter a URL, and then pressing enter to load that page. With the concepts learned above, we can now do that entirely using PyAutoGUI. After pressing run on the cell below, sit back and relax while it opens the PyAutoGUI docs for you!